+++ font/opentype/cidcff.go
@@ -56,7 +56,7 @@ func (f *embeddedCFFComposite) DecodeWidth(s pdf.String) (float64, int) {
 	for code, cid := range f.AllCIDs(s) {
 		gid := f.GID(cid)
 		// TODO(voss): deal with different Font Matrices for different private dicts.
-		width := float64(f.sfnt.GlyphWidth(gid)) * f.sfnt.FontMatrix[0]
+		width := f.sfnt.GlyphWidth(gid) * f.sfnt.FontMatrix[0]
 		return width, len(code)
 	}
 	return 0, 0
@@ -64,7 +64,7 @@ func (f *embeddedCFFComposite) DecodeWidth(s pdf.String) (float64, int) {
 
 func (f *embeddedCFFComposite) AppendEncoded(s pdf.String, gid glyph.ID, text string) (pdf.String, float64) {
 	// TODO(voss): deal with different Font Matrices for different private dicts.
-	width := float64(f.sfnt.GlyphWidth(gid)) * f.sfnt.FontMatrix[0]
+	width := f.sfnt.GlyphWidth(gid) * f.sfnt.FontMatrix[0]
 	s = f.CIDEncoder.AppendEncoded(s, gid, text)
 	return s, width
 }
