+++ font/truetype/cid.go
@@ -53,14 +53,14 @@ func (f *embeddedComposite) WritingMode() font.WritingMode {
 func (f *embeddedComposite) DecodeWidth(s pdf.String) (float64, int) {
 	for code, cid := range f.AllCIDs(s) {
 		gid := f.GID(cid)
-		width := float64(f.sfnt.GlyphWidth(gid)) / float64(f.sfnt.UnitsPerEm)
+		width := f.sfnt.GlyphWidth(gid) / float64(f.sfnt.UnitsPerEm)
 		return width, len(code)
 	}
 	return 0, 0
 }
 
 func (f *embeddedComposite) AppendEncoded(s pdf.String, gid glyph.ID, text string) (pdf.String, float64) {
-	width := float64(f.sfnt.GlyphWidth(gid)) / float64(f.sfnt.UnitsPerEm)
+	width := f.sfnt.GlyphWidth(gid) / float64(f.sfnt.UnitsPerEm)
 	s = f.CIDEncoder.AppendEncoded(s, gid, text)
 	return s, width
 }
